added 'updated' flag to improve dependency tracking
We were missing needed recomputes for second order
dependency changes.  For example:

    a <- 1
    b <- f(a)
    c <- g(b)

If in the next run, the first line is changed to 'a <- 2', then we
weren't recomputing c because it depends on b and the hash there does
change.

The solution is to add an updated flag so that even when the hash 
doesn't change, we detect that the value might have and recompute...


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/weaver@19685 bc3139a8-67e5-0310-9ffc-ced21a209358